+++ apps/evals/src/main.js
@@ -440,3 +440,4 @@ main().catch((e) => {
   process.exitCode = 1;
 });
 
+