+++ core/dialogs/todo_bot_admin.js
@@ -1170,3 +1170,4 @@ function registerAdminCommands({
 
 module.exports = { registerAdminCommands };
 
+