+++ core/runtime/healthcheck_lib.js
@@ -201,3 +201,4 @@ async function runHealthcheck({ wantPostgres = true, wantNotion = true, wantTele
 
 module.exports = { runHealthcheck, checkPostgres, checkNotion, checkTelegramSend, tailId };
 
+