+++ infra/deploy/prod_deploy.sh
@@ -84,3 +84,4 @@ fi
 
 echo "Deploy done"
 
+